+++ stroom-config/stroom-config-global-impl/src/test/java/stroom/config/global/impl/TestConfigMapper.java
@@ -4,21 +4,28 @@ import stroom.config.app.AppConfig;
 import stroom.config.global.shared.ConfigProperty;
 import stroom.config.global.shared.OverrideValue;
 import stroom.docref.DocRef;
+import stroom.util.config.PropertyUtil.Prop;
 import stroom.util.io.ByteSize;
+import stroom.util.logging.AsciiTable;
+import stroom.util.logging.AsciiTable.Column;
 import stroom.util.logging.LogUtil;
 import stroom.util.shared.AbstractConfig;
 import stroom.util.shared.PropertyPath;
 import stroom.util.time.StroomDuration;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.google.common.reflect.TypeToken;
 import io.dropwizard.Configuration;
 import io.dropwizard.configuration.ConfigurationException;
+import io.vavr.Tuple;
+import io.vavr.Tuple2;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
+import java.lang.reflect.Type;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -28,6 +35,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
+import java.util.function.BiFunction;
 import java.util.function.Function;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
@@ -65,18 +73,23 @@ class TestConfigMapper {
 
     @Test
     void getDataTypeNames() {
-        TestConfig appConfig = new TestConfig();
-        ConfigMapper configMapper = new ConfigMapper(appConfig);
+        final TestConfig appConfig = new TestConfig();
+        final ConfigMapper configMapper = new ConfigMapper(appConfig);
 
-        Collection<ConfigProperty> configProperties = configMapper.getGlobalProperties();
+        final Collection<ConfigProperty> configProperties = configMapper.getGlobalProperties();
 
-        String txt = configProperties.stream()
+        final List<Tuple2<String, String>> rows = configProperties.stream()
                 .sorted(Comparator.comparing(ConfigProperty::getName))
                 .map(configProperty ->
-                        LogUtil.message("{}, {}", configProperty.getName(), configProperty.getDataTypeName()))
-                .collect(Collectors.joining("\n"));
+                        Tuple.of(configProperty.getName().toString(), configProperty.getDataTypeName()))
+                .collect(Collectors.toList());
 
-        LOGGER.debug("Properties\n{}", txt);
+        final String asciiTable = AsciiTable.builder(rows)
+                .withColumn(Column.of("Property Path", Tuple2::_1))
+                .withColumn(Column.of("Data Type", Tuple2::_2))
+                .build();
+
+        LOGGER.debug("Properties\n{}", asciiTable);
     }
 
     @Test
@@ -307,11 +320,21 @@ class TestConfigMapper {
                 "1MiB",
                 ByteSize::parse);
 
-        doUpdateValueTest(
+        doUpdateValueTestComplex(
                 "stroom.docRefProp",
                 TestConfig::getDocRefProp,
                 ",docRef(aaaaaa,bbbbbbb,ccccccc)",
-                str -> ConfigMapper.convertToObject(str, DocRef.class));
+                (prop, str) -> ConfigMapper.convertToObject(prop, str, DocRef.class));
+
+        final Type stateListType = new TypeToken<List<TestConfig.State>>() {
+        }.getType();
+
+        doUpdateValueTestComplex(
+                "stroom.stateListProp",
+                TestConfig::getStateListProp,
+                ",ON,OFF",
+                (prop, str) ->
+                        ConfigMapper.convertToObject(prop, str, stateListType));
 
         doUpdateValueTest(
                 "stroom.stateProp",
@@ -324,6 +347,17 @@ class TestConfigMapper {
                                final Function<TestConfig, T> getter,
                                final String newValueAsStr,
                                final Function<String, T> parseFunc) {
+        doUpdateValueTestComplex(
+                path,
+                getter,
+                newValueAsStr,
+                (prop, s) -> parseFunc.apply(s));
+    }
+
+    <T> void doUpdateValueTestComplex(final String path,
+                                      final Function<TestConfig, T> getter,
+                                      final String newValueAsStr,
+                                      final BiFunction<Prop, String, T> parseFunc) {
 
         LOGGER.info("Testing {}, with new value {}", path, newValueAsStr);
 
@@ -334,6 +368,9 @@ class TestConfigMapper {
         final PropertyPath fullPath = PropertyPath.fromPathString(path);
 
         final ConfigMapper configMapper = new ConfigMapper(testConfig);
+
+        final Prop prop = configMapper.getProp(fullPath)
+                .orElseThrow();
 
         boolean isValidPath = configMapper.validatePropertyPath(PropertyPath.fromPathString(path));
 
@@ -348,7 +385,7 @@ class TestConfigMapper {
         configProperty.setDatabaseOverrideValue(newValueAsStr);
         configMapper.decorateDbConfigProperty(configProperty);
 
-        final T newObj = parseFunc.apply(newValueAsStr);
+        final T newObj = parseFunc.apply(prop, newValueAsStr);
 
         LOGGER.info("{} - {} => {}", newObj.getClass().getSimpleName(), originalObj, newObj);